+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features / Enhancements
 
 - Updated Autosize Code Editor toolbar (#362)
+- Added helper statusColor from specific field (#375)
 
 ## 5.4.0 (2024-09-12)