+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Updated Autosize Code Editor toolbar (#362)
 - Added helper statusColor from specific field (#375)
+- Updated E2E tests (#377)
 
 ## 5.4.0 (2024-09-12)