+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added helper statusColor from specific field (#375)
 - Updated E2E tests (#377)
 - Updated to Grafana 11.4 and dependencies (#378)
+- Added replace variables in URLs (#376)
 
 ## 5.4.0 (2024-09-12)